stage0/stage1 app: unify filenames
Currently we have mixture of file and directory conventions for app subcommands in
stage0 and stage1.

This fixes it by consolidating them following Go standards (underscore
delimiters in filenames).
