+++ stage1/aci/aci-manifest.in
@@ -35,19 +35,19 @@
         },
         {
             "name": "coreos.com/rkt/stage1/app/add",
-            "value": "/app-add"
+            "value": "/app_add"
         },
         {
             "name": "coreos.com/rkt/stage1/app/rm",
-            "value": "/app-rm"
+            "value": "/app_rm"
         },
         {
             "name": "coreos.com/rkt/stage1/app/start",
-            "value": "/app-start"
+            "value": "/app_start"
         },
         {
             "name": "coreos.com/rkt/stage1/app/stop",
-            "value": "/app-stop"
+            "value": "/app_stop"
         },
         {
             "name": "coreos.com/rkt/stage1/interface-version",
