+++ stage1/app_rm/app_rm.go
@@ -132,7 +132,7 @@ func cleanupStage0(appName *types.ACName, enterCmd []string) error {
 	// last cleaning steps are performed after entering pod context
 	tasks := [][]string{
 		// inception: call itself to clean stage1 before proceeding
-		{"/app-rm", "--stage=1", fmt.Sprintf("--app=%s", appName), fmt.Sprintf("--debug=%t", debug)},
+		{"/app_rm", "--stage=1", fmt.Sprintf("--app=%s", appName), fmt.Sprintf("--debug=%t", debug)},
 		// all cleaned-up, let systemd reload and forget about this app
 		{"/usr/bin/systemctl", "daemon-reload"},
 	}
